// vscode-extension/client/src/plcopen/webview/openplc-ladder/atoms/utils/types.ts
export type BlockVariant = {
  name: string
  type: string
  variables: { id?: string; name: string; class: string; type: { definition: string; value: string } }[]
  documentation: string
  extensible: boolean
}